// src/features/editableProfileCard/model/consts/consts.ts
export enum ValidateProfileError {
	INCORRECT_USER_DATA = 'INCORRECT_USER_DATA',
	INCORRECT_AGE = 'INCORRECT_AGE',
	INCORRECT_CITY = 'INCORRECT_CITY',
	INCORRECT_USERNAME = 'INCORRECT_USERNAME',
	INCORRECT_COUNTRY = 'INCORRECT_COUNTRY',
	INCORRECT_CURRENCY = 'INCORRECT_CURRENCY',
	NO_DATA = 'NO_DATA',
	SERVER_ERROR = 'SERVER_ERROR',
}